Add files via upload
Program reads information from a .txt file, and stores it in 'Rating' objects.
Format of file is: [<User ID> <Item ID> <Rating>].
Program stores Ratings with similar 'user id' alongside each other in a (B)inary (S)earch (T)ree that uses their 'item id' as a key, and stores that BST in a BST that uses the first BST's Ratings' 'user id' as a key. The same process is duplicated vice versa, such that we have two big BSTs.
The main class is 'RatingManager'.
The program can extract the average of ratings given by a user, the average rating given to an item, the highest rated item (items if they share the same rating).
Finally, the program can estimate the rating a user might give to an item by checking the average rating his neighbors gave to it. A user's neighbors are the users that gave items he rated similar ratings.